refactor: move lines to `Cell.hs`
