+++ src/Dungeon/Map/Tile.hs
@@ -1,91 +1,23 @@
-{-# LANGUAGE DeriveGeneric   #-}
-{-# LANGUAGE TemplateHaskell #-}
+{-# LANGUAGE DeriveGeneric #-}
 
 module Dungeon.Map.Tile
-    ( CellMap
-    , cellMap
-    , widthAndHeight
-    , changeTileAt
-    , walkableMap
-    , transparentMap
-    , tileIdAt
-    , isWalkableAt
-    , Tile
+    ( Tile
     , TileCollection
     , TileId
     , tile
-    , allWallTiles
     , isWalkable
     , isTransparent
+    , wallTile
     , floorTile
     , upStairs
     , downStairs
     ) where
 
-import           Actor            (Actor)
-import           Control.Lens     (makeLenses, (&), (.~), (^.))
-import           Coord            (Coord)
-import           Data.Array       (Array, bounds, (!), (//))
-import           Data.Binary      (Binary)
-import           Data.Maybe       (isJust)
-import qualified Dungeon.Map      as M
-import           Dungeon.Map.Bool (BoolMap)
-import           GHC.Generics     (Generic)
-import           Linear.V2        (V2 (V2))
+import           Data.Array   (Array)
+import           Data.Binary  (Binary)
+import           GHC.Generics (Generic)
 
 type TileId = Int
-
-data Cell =
-    Cell
-        { _tileId :: TileId
-        , actor   :: Maybe Actor
-        }
-    deriving (Show, Ord, Eq, Generic)
-
-makeLenses ''Cell
-
-instance Binary Cell
-
-newtype CellMap =
-    CellMap (Array (V2 Int) Cell)
-    deriving (Show, Ord, Eq, Generic)
-
-instance Binary CellMap
-
-cellMap :: Array (V2 Int) TileId -> CellMap
-cellMap = CellMap . fmap (`Cell` Nothing)
-
-widthAndHeight :: CellMap -> V2 Int
-widthAndHeight (CellMap m) = snd (bounds m) + V2 1 1
-
-changeTileAt :: Coord -> TileId -> CellMap -> Maybe CellMap
-changeTileAt c i (CellMap m)
-    | isJust $ tileIdAt c (CellMap m) = Just $ CellMap $ m // [(c, newTile)]
-    | otherwise = Nothing
-  where
-    newTile = m ! c & tileId .~ i
-
-walkableMap :: TileCollection -> CellMap -> BoolMap
-walkableMap tc (CellMap m) = isWalkable . ((tc !) . (^. tileId)) <$> m
-
-transparentMap :: TileCollection -> CellMap -> BoolMap
-transparentMap tc (CellMap m) = isTransparent . ((tc !) . (^. tileId)) <$> m
-
-tileIdAt :: Coord -> CellMap -> Maybe TileId
-tileIdAt c t = (^. tileId) <$> cellAt c t
-
-cellAt :: Coord -> CellMap -> Maybe Cell
-cellAt c (CellMap m)
-    | c >= lower && c <= upper = Just $ m ! c
-    | otherwise = Nothing
-  where
-    (lower, upper) = bounds m
-
-isWalkableAt :: Coord -> TileCollection -> CellMap -> Bool
-isWalkableAt c tc t =
-    case tileIdAt c t of
-        Just x  -> isWalkable (tc ! x)
-        Nothing -> False
 
 data Tile =
     Tile
@@ -100,9 +32,6 @@ type TileCollection = Array Int Tile
 
 tile :: Bool -> Bool -> Tile
 tile = Tile
-
-allWallTiles :: V2 Int -> CellMap
-allWallTiles wh = CellMap $ M.generate wh (const (Cell wallTile Nothing))
 
 isWalkable :: Tile -> Bool
 isWalkable = walkable
